Updated the mission statement on the About page.
Made changes as listed on figma to mission statement on the About page. Also changed the title on the about page from "About Us" to "More About Us" as shown on figma. Changed the devdog logo image to the demo image as shown on figma. Changed the font size under p (paragraph) from 1.2rem to 1.5rem on the globals.css sheet so the text better fits the webpage. Also changed changed the background-size on the globals.css sheet from 24% to 35% to better match the design shown on figma.