Add basic DevDogs Academy page
RFC. Thought I'd try my hand at a little more frontend stuff. Everything
is best-effort, sorry if this has silly mistakes. There is still plenty
left to polish, but this should provide a passable base to build off of.

To be clear, this should not go live as-is (don't link to it from
anywhere). It's still a work-in-progress page.